Rearrange test code for exhaustiveness
