+++ test/ClangImporter/enum-dataflow.swift
@@ -1,4 +1,4 @@
-// RUN: %target-swift-frontend(mock-sdk: %clang-importer-sdk) -emit-sil %s -verify
+// RUN: %target-swift-frontend(mock-sdk: %clang-importer-sdk) -typecheck %s -verify
 
 // REQUIRES: objc_interop
 
@@ -7,12 +7,15 @@ import user_objc
 
 let aliasOriginal = NSAliasesEnum.byName
 
-switch aliasOriginal { // expected-error {{switch must be exhaustive, consider adding missing cases}}
+switch aliasOriginal { // expected-error {{switch must be exhaustive, consider adding missing cases:}}
+// expected-note@-1 {{missing case: '.differentValue'}}
 case .original:
   break
 }
 
-switch aliasOriginal { // expected-error {{switch must be exhaustive, consider adding missing cases}}
+switch aliasOriginal { // expected-error {{switch must be exhaustive, consider adding missing cases:}}
+// expected-note@-1 {{missing case: '.original'}}
+// expected-note@-2 {{missing case: '.differentValue'}}
 case .bySameValue:
   break
 }
