+++ test/ClangImporter/enum-error.swift
@@ -84,7 +84,7 @@ func testError() {
 
   switch (terr) { case .TENone, .TEOne: break }
   // EXHAUSTIVE: [[@LINE-1]]:{{.+}}: error: switch must be exhaustive, consider adding missing cases
-
+  // EXHAUSTIVE: [[@LINE-2]]:{{.+}}: note: missing case: '.TETwo'
   let _ = TestError.Code(rawValue: 2)!
 
   do {
