+++ test/Constraints/patterns.swift
@@ -16,6 +16,8 @@ case (0..<10, _, _),
      (0...9, _, _),
      (0...9, 2.5, "three"):
   ()
+default:
+  ()
 }
 
 switch (1, 2) {
@@ -55,6 +57,8 @@ case is B,
   ()
 case is E:
   ()
+default:
+  ()
 }
 
 switch bp {
@@ -67,6 +71,8 @@ case let s as S:
   s.s()
 case let e as E:
   e.e()
+default:
+  ()
 }
 
 // Super-to-subclass.
@@ -75,6 +81,8 @@ switch db {
 case is D:
   ()
 case is E: // expected-warning {{always fails}}
+  ()
+default:
   ()
 }
 
@@ -175,6 +183,7 @@ func ~= <T : Equatable>(lhs: T?, rhs: T?) -> Bool {
 
 switch 4 as Int? {
 case x?.method(): break // match value
+default: break
 }
 
 switch 4 {
