+++ test/FixCode/fixits-switch.swift.result
@@ -1,4 +1,4 @@
-// RUN: not %swift -emit-sil -target %target-triple %s -emit-fixits-path %t.remap -I %S/Inputs
+// RUN: not %swift -emit-sil -target %target-triple %s -emit-fixits-path %t.remap -I %S/Inputs -diagnostics-editor-mode
 // RUN: c-arcmt-test %t.remap | arcmt-test -verify-transformed-files %s.result
 
 enum E1 : Int {
@@ -50,13 +50,13 @@ func foo4(_ e : E2) -> Int {
   switch e {
   case .e2:
     return 1
-  case .e1(let a, let s): <#code#>
-case .e3(let a): <#code#>
+  case .e1(_, _): <#code#>
+case .e3(_): <#code#>
 case .e4(_): <#code#>
 case .e5(_, _): <#code#>
-case .e6(let a, _): <#code#>
+case .e6(_, _): <#code#>
 case .e7: <#code#>
-case .e8(let a, _, _): <#code#>
+case .e8(_, _, _): <#code#>
 case .e9(_, _, _): <#code#>
 }
 }
@@ -65,7 +65,10 @@ func foo5(_ e : E1) -> Int {
   switch e {
   case _ where e.rawValue > 0:
     return 1
-  default: <#code#>
+  case .e1: <#code#>
+case .e2: <#code#>
+case .e3: <#code#>
+case .e4: <#code#>
 }
 }
 
@@ -73,13 +76,13 @@ func foo6(_ e : E2) -> Int {
   switch e {
   case let .e1(x, y):
     return x + y
-  case .e2(let a): <#code#>
-case .e3(let a): <#code#>
+  case .e2(_): <#code#>
+case .e3(_): <#code#>
 case .e4(_): <#code#>
 case .e5(_, _): <#code#>
-case .e6(let a, _): <#code#>
+case .e6(_, _): <#code#>
 case .e7: <#code#>
-case .e8(let a, _, _): <#code#>
+case .e8(_, _, _): <#code#>
 case .e9(_, _, _): <#code#>
 }
 }
@@ -89,11 +92,12 @@ func foo7(_ e : E2) -> Int {
   case .e2(1): return 0
   case .e1: return 0
   case .e3: return 0
-  case .e4(_): <#code#>
+  case .e2(_): <#code#>
+case .e4(_): <#code#>
 case .e5(_, _): <#code#>
-case .e6(let a, _): <#code#>
+case .e6(_, _): <#code#>
 case .e7: <#code#>
-case .e8(let a, _, _): <#code#>
+case .e8(_, _, _): <#code#>
 case .e9(_, _, _): <#code#>
 }
 }
