+++ test/Parse/invalid.swift
@@ -51,7 +51,7 @@ func testNotCoveredCase(x: Int) {
     break
   }
 
-  switch x { // expected-error{{'switch' statement body must have at least one 'case' or 'default' block}}
+  switch x { // expected-error{{'switch' statement body must have at least one 'case' or 'default' block; do you want to add a default case?}}
 #if true // expected-error {{all statements inside a switch must be covered by a 'case' or 'default'}}
   case 1:
     break
