+++ test/Parse/matching_patterns.swift
@@ -55,7 +55,7 @@ case _:
 
 var e : Any = 0
 
-switch e {
+switch e { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 // 'is' pattern.
 case is Int,
      is A<Int>,
@@ -149,7 +149,9 @@ struct ContainsEnum {
   }
 
   func member(_ n: Possible<Int>) {
-    switch n {
+    switch n { // expected-error {{switch must be exhaustive, consider adding missing cases:}}
+    // expected-note@-1 {{missing case: '.Mere(_)'}}
+    // expected-note@-2 {{missing case: '.Twain(_, _)'}}
     case ContainsEnum.Possible<Int>.Naught,
          ContainsEnum.Possible.Naught,
          Possible<Int>.Naught,
@@ -161,7 +163,9 @@ struct ContainsEnum {
 }
 
 func nonmemberAccessesMemberType(_ n: ContainsEnum.Possible<Int>) {
-  switch n {
+  switch n { // expected-error {{switch must be exhaustive, consider adding missing cases:}}
+  // expected-note@-1 {{missing case: '.Mere(_)'}}
+  // expected-note@-2 {{missing case: '.Twain(_, _)'}}
   case ContainsEnum.Possible<Int>.Naught,
        .Naught:
     ()
