+++ test/Parse/switch.swift
@@ -71,7 +71,7 @@ default:
 }
 
 // Multiple cases per case block
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 case 0: // expected-error {{'case' label in a 'switch' should have at least one executable statement}} {{8-8= break}}
 case 1:
   x = 0
@@ -83,7 +83,7 @@ default:
   x = 0
 }
 
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 case 0:
   x = 0
 case 1: // expected-error {{'case' label in a 'switch' should have at least one executable statement}} {{8-8= break}}
@@ -95,7 +95,7 @@ case 0:
 default: // expected-error {{'default' label in a 'switch' should have at least one executable statement}} {{9-9= break}}
 }
 
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 case 0:
   ; // expected-error {{';' statements are not allowed}} {{3-5=}}
 case 1:
@@ -142,22 +142,22 @@ default: // expected-error{{additional 'case' blocks cannot appear after the 'de
   x = 0
 }
 
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 default where x == 0: // expected-error{{'default' cannot be used with a 'where' guard expression}}
   x = 0
 }
 
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 case 0: // expected-error {{'case' label in a 'switch' should have at least one executable statement}} {{8-8= break}}
 }
 
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 case 0: // expected-error{{'case' label in a 'switch' should have at least one executable statement}} {{8-8= break}}
 case 1:
   x = 0
 }
 
-switch x {
+switch x { // expected-error {{switch must be exhaustive, consider adding a default clause:}}
 case 0:
   x = 0
 case 1: // expected-error{{'case' label in a 'switch' should have at least one executable statement}} {{8-8= break}}
@@ -256,7 +256,7 @@ case (var a, var b): // expected-warning {{variable 'a' was never mutated; consi
 }
 
 func test_label(x : Int) {
-Gronk:
+Gronk: // expected-error {{switch must be exhaustive, consider adding a default clause:}}
   switch x {
   case 42: return
   }
