+++ test/SILOptimizer/unreachable_code.swift
@@ -129,23 +129,6 @@ func testSwitchEnum(_ xi: Int) -> Int {
     x += 1
   }
 
-  switch cond { // no warning
-  case .Two: 
-    x += 1
-  }
-
-  switch cond { // expected-error{{switch must be exhaustive}}
-  case .One:
-    x += 1
-  }
-
-  switch cond { // expected-error{{switch must be exhaustive}}
-  case .One:
-    x += 1
-  case .Three:
-    x += 1
-  }
-
   switch cond { // expected-warning{{switch condition evaluates to a constant}}
   case .Two: 
     x += 1
@@ -162,110 +145,6 @@ func testSwitchEnum(_ xi: Int) -> Int {
   
   return x
 }
-
-
-// Treat nil as .none and do not emit false 
-// non-exhaustive warning.
-func testSwitchEnumOptionalNil(_ x: Int?) -> Int {
-  switch x { // no warning
-  case .some(_):
-    return 1
-  case nil:
-    return -1
-  }
-}
-
-// Do not emit false non-exhaustive warnings if both
-// true and false are covered by the switch.
-func testSwitchEnumBool(_ b: Bool, xi: Int) -> Int {
-  var x = xi
-  let Cond = b
-  
-  switch Cond { // no warning
-  default:
-    x += 1
-  }
-
-  switch Cond { // expected-error{{switch must be exhaustive}}
-  case true:
-    x += 1
-  }
-
-  switch Cond { // expected-error{{switch must be exhaustive}}
-  case false:
-    x += 1
-  }
-
-  switch Cond { // no warning
-  case true:
-    x += 1
-  case false:
-    x -= 1
-  }
-
-  return x
-}
-
-func testSwitchOptionalBool(_ b: Bool?, xi: Int) -> Int {
-  var x = xi
-  switch b { // No warning
-  case .some(true):
-    x += 1
-  case .some(false):
-    x += 1
-  case .none:
-    x -= 1
-  }
-
-  switch b { // expected-error{{switch must be exhaustive}}
-  case .some(true):
-    x += 1
-  case .none:
-    x -= 1
-  }
-
-  return xi
-}
-
-// Do not emit false non-exhaustive warnings if both 
-// true and false are covered for a boolean element of a tuple.
-func testSwitchEnumBoolTuple(_ b1: Bool, b2: Bool, xi: Int) -> Int {
-  var x = xi
-  let Cond = (b1, b2)
-  
-  switch Cond { // no warning
-  default:
-    x += 1
-  }
-
-  switch Cond { // expected-error{{switch must be exhaustive}} expected-error{{switch must be exhaustive}}
-  case (true, true):
-    x += 1
-    // FIXME: Two expect statements are written, because unreachable diagnostics produces N errors
-    // for non-exhaustive switches on tuples of N elements
-  }
-
-  switch Cond { // expected-error{{switch must be exhaustive}} expected-error{{switch must be exhaustive}}
-  case (false, true):
-    x += 1
-    // FIXME: Two expect statements are written, because unreachable diagnostics produces N errors
-    // for non-exhaustive switches on tuples of N elements
-  }
-
-  switch Cond { // no warning
-  case (true, true):
-    x += 1
-  case (true, false):
-    x += 1
-  case (false, true):
-    x -= 1
-  case (false, false):
-    x -= 1
-  }
-
-  return x
-}
-
 
 @_silgen_name("exit") func exit() -> Never
