+++ test/stmt/switch_stmt1.swift
@@ -6,6 +6,8 @@ enum E {
 }
 
 func foo1(e : E, i : Int) {
-  switch e {} // expected-error{{'switch' statement body must have at least one 'case' or 'default' block; do you want to add missing cases?}}{{13-13=case .e1: <#code#>\ncase .e2: <#code#>\n}}
+  switch e {} // expected-error{{switch must be exhaustive, consider adding missing cases:}}
+  // expected-note@-1 {{missing case: '.e1'}}
+  // expected-note@-2 {{missing case: '.e2'}}
   switch i {} // expected-error{{'switch' statement body must have at least one 'case' or 'default' block; do you want to add a default case?}}{{13-13=default: <#code#>\n}}
 }
